requirements.txt
numpy==1.24.4
opencv-python-headless==4.8.1.78
captcha==0.5.0
pillow==10.0.1
tensorflow[and-cuda]
tqdm
pandas
requests
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6